requirements.txt
streamlit
torch
transformers
datasets[audio]
openai
supabase
python-dotenv
librosa
soundfile
evaluate
jiwer
numpy
pandas
pydantic
tensorboard
gradio
spacy
pytextrank
yake
bert-extractive-summarizer
sentencepiece
rouge-score==0.1.2
nltk==3.8.1
sentence-transformers==2.2.2
bert-score==0.3.13
plotly==5.18.0
